Successfully added redux to application
