+++ src/components/AddBook.js
@@ -1,33 +1,19 @@
-import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { addBook } from '../redux/books/books';
+import React from 'react';
+import PropTypes from 'prop-types';
 
-const AddBook = () => {
-  const [title, setTitle] = useState('');
-  const [author, setAuthor] = useState('');
-  const dispatch = useDispatch();
-  const submitBook = (e) => {
-    e.preventDefault();
-    const form = e.target;
-    const book = {
-      id: new Date().getTime().toString(),
-      title,
-      author,
-    };
-    dispatch(addBook(book));
-    setTitle('');
-    setAuthor('');
-    form.reset();
-  };
-  return (
-    <>
-      <form id="form" onSubmit={submitBook}>
-        <input type="text" id="book-title" name="title" className="book" placeholder="Book Name..." onChange={(e) => setTitle(e.target.value)} required />
-        <input type="text" id="book-author" name="author" className="book" placeholder="Book Author..." onChange={(e) => setAuthor(e.target.value)} required />
-        <input type="submit" />
-      </form>
-    </>
-  );
+const AddBook = ({ submitBook, bookDetail }) => (
+  <>
+    <form id="form" onSubmit={submitBook}>
+      <input type="text" id="book-title" name="title" className="book" placeholder="Book Name..." onChange={bookDetail} required />
+      <input type="text" id="book-author" name="author" className="book" placeholder="Book Author..." onChange={bookDetail} required />
+      <input type="submit" />
+    </form>
+  </>
+);
+
+AddBook.propTypes = {
+  submitBook: PropTypes.func.isRequired,
+  bookDetail: PropTypes.func.isRequired,
 };
 
 export default AddBook;
